Added ability to download development and distribution certificates.  Added comments
This is a useful feature if you use Jenkins to product AdHoc builds.  If a profile name is specified and it is found in the development profiles, the process is exited after it is downloaded.  Next, the distribution profiles are checked for a match.  If no profile is specified, all development and distribution profiles are downloaded